requirements.txt
streamlit
requests
PyPDF2
python-dotenv
huggingface_hub